Refactor: Rename likes to highs in Topic entity
Replace "likes" with "highs" to better reflect the
intended meaning of user interaction with topics.  This
change affects the entity, DTOs, mapper, controller, and
test classes.  The `LikeEntity` is removed.

+++ src/main/java/br/com/soupaulodev/forumhub/modules/topic/controller/dto/TopicDetailsResponseDTO.java
@@ -35,8 +35,8 @@ public record TopicDetailsResponseDTO(
         UUID forumId,
         UUID creatorId,
         String creatorUsername,
-        int likeCount,
-        int commentCount,
+        Long likeCount,
+        Long commentCount,
         List<CommentResponseDTO> comments,
         Instant createdAt,
         Instant updatedAt
